Wrapper handles more errors, added debugging flags, new AST node for return statement, parser updates to correctly parse empty_main.

+++ fc.py
@@ -2,6 +2,7 @@
 #Fork compiler toolchain script
 import argparse
 import re, os
+from sys import exit
 
 def main():
   #Parse arguments
@@ -10,11 +11,19 @@ def main():
   regex_delete = re.compile("(^\s*//.*)|(^\s*$)")
   args = parser.parse_args()
   files = args.files
+  #Check that parser exists
+  if not os.path.exists("./parser"):
+    print("Parser binary not found in current directory.")
+    exit(1)
   #Preprocessing
   temp_files = [x + '.wrapper_tmp_file' for x in files]
   for file_in,file_out in zip(files,temp_files):
-    f_in = open(file_in,'r')
-    f_out = open(file_out,'w')
+    try:
+      f_in = open(file_in,'r')
+      f_out = open(file_out,'w')
+    except (FileNotFoundError,PermissionError) as e:
+      print(e)
+      exit(2)
     for line in f_in:
       if not regex_delete.match(line):
         f_out.write(line)
@@ -22,7 +31,7 @@ def main():
     f_out.close()
   #Call parser on temp_files
   for file in temp_files:
-    os.system("./parser {}".format(file))
+    os.system("valgrind ./parser {}".format(file))
   #Postprocessing
   for file in temp_files:
     os.remove(file)
